chore: add new package for high-level multipart support
