+++ packages/multipart/README.md
@@ -1,0 +1,5 @@
+
+# HTTP support for SenseJS
+
+This package provides HTTP support to SENSE.js framework based on koa.
+
