+++ packages/multipart/jest.config.cjs
@@ -1,0 +1,2 @@
+
+module.exports = {...require('../../jest.base-config.cjs')};
