+++ packages/multipart/src/error.ts
@@ -1,0 +1,10 @@
+export class MultipartError implements Error {
+  readonly name;
+  constructor(readonly message: string) {
+    this.name = new.target.name;
+  }
+}
+
+export class MultipartLimitExceededError extends MultipartError {}
+
+export class InvalidMultipartBodyError extends MultipartError {}
