+++ packages/multipart/tsconfig.build-cjs.json
@@ -1,0 +1,9 @@
+{
+  "extends": "./src/tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "module": "CommonJS",
+    "outDir": "dist-cjs/lib",
+    "paths": {}
+  }
+}
