+++ packages/multipart/tsconfig.test.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.test",
+  "include": [
+    "tests/**/*.ts",
+    "src/**/*.ts"
+  ]
+}
